Moved database related stuff to a separate file
